Gate the admin dashboard server-side, not client-side
The dashboard layout only checked auth via ProtectedRoute, a client
component: it rendered a full skeleton shell for every visitor
before an effect ran useAuth() and redirected if unauthenticated.
The server never gated the page at all - it just shipped a fake
loading UI to anyone who asked, authenticated or not, before
possibly bouncing them client-side. That's the opposite of
"deny by default, server-side" (this project's own stated rule),
even though every underlying API route was already independently
protected.

The dashboard layout is now an async Server Component that checks
getCurrentSession() and redirects before anything renders - no
skeleton flash, no client-side round trip, nothing sent to an
unauthenticated request but the redirect itself. The interactive
shell (theme, sidebar, search) moved into DashboardShell, a plain
client component with no gating responsibility. ProtectedRoute and
DashboardSkeleton had no other callers and are deleted.

Verified against the real dev server: no cookie -> 307 to
/auth/sign-in with no dashboard markup in the response; valid
session -> 200 with the real dashboard.
